Add production default url and rename jwt_auth.rb
