+++ app/services/jwt_auth_service.rb
